agentforce/testing/requirements.txt
# Requirements for Agentforce testing scripts
requests>=2.31.0
python-dotenv>=1.0.0
